# oas/OAS.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant Salon API
  description: |
    REST API for the multi-tenant SaaS salon/salon management platform.

    ## Module / Access Hierarchy

    The API is split into four distinct path namespaces:

    | Namespace | Base Path | Who uses it |
    |---|---|---|
    | **Salon Onboarding** | `/api/salon-onboarding` | Anyone registering a new salon tenant; platform-level listing |
    | **Customer** | `/api/salon/{salonId}/...` | Public/customer-facing — browse, book, read |
    | **Admin** | `/api/salon-admin/{salonId}/...` | Salon owner/administrator — manage, configure, operate. Also callable by a platform super-admin for any salon, since super-admin management of a specific salon reuses these endpoints rather than duplicating them. |
    | **Super Admin** | `/api/salon-super-admin/...` | Platform super-admin — cross-tenant management of all salons |
    | **Staff Portal** | `/api/salon-staff/...` | Authenticated staff member — self-service profile, appointments, personal holidays |
    | **Utility** | `/api/salon-utility/...` | Any consumer needing reference data (countries with embedded currency info) |

    ### Customer sub-paths
    - `/api/salon/{salonId}/services/...` — browse the service catalog
    - `/api/salon/{salonId}/staff/...` — view staff members
    - `/api/salon/{salonId}/booking/...` — slot discovery and appointment creation
    - `/api/salon/{salonId}/website` — public website theme
...

    ### Admin sub-paths
...

...
servers:
...
paths:
...
  /api/salon/{salonId}/booking/slots:
    parameters:
      - $ref: "#/components/parameters/SalonId"
    get:
      tags: [Booking]
      summary: Get booking slots for a service on a given date
      description: |
        Customer-facing. Returns all time slots for a service on a specific date — both
        available and already-booked ones. Slots are derived from each eligible staff
        member's weekly schedule or date-specific override.
        Each slot carries a `booked` flag: `false` means the slot is open and selectable;
        `true` means it is taken (existing non-cancelled booking) and should be shown
        grayed-out / non-selectable in the UI.
        Optionally filter to a single staff member via `staffId`.
        Returns an **empty array** when the date falls within a salon closure.
      operationId: getAvailableSlots
      parameters:
        - name: serviceId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2026-08-01"
        - name: staffId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter slots for a specific staff member; omit to see slots across all available staff
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailableSlot"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/salon/{salonId}/booking:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    post:
      tags: [Booking]
      summary: Create a booking
      description: |
        Customer-facing. Books an available slot for a service. `staffId` is optional — the
        system auto-assigns the first available staff member if omitted.
        Returns HTTP 400 when `appointmentDate` falls within a salon closure.
        The initial booking status depends on the salon's `bookingRequiresConfirmation` setting:
        - `false` (default) → status is **CONFIRMED** immediately.
        - `true` → status is **PENDING**; an admin must confirm before the appointment is active.
        Emits **BookingCreatedEvent** → customer receives a booking confirmation notification.
        Also emits **StaffBookingAssignedEvent** → the assigned staff member is notified.
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Salon is closed on the requested date (salon closure)
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Requested slot is no longer available
  /api/salon/{salonId}/booking/{bookingId}:
    parameters:
      - $ref: "#/components/parameters/SalonId"
      - $ref: "#/components/parameters/BookingId"

    get:
      tags: [Booking]
      summary: Get a booking by ID
      operationId: getBooking
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Customer — Website Theme ──────────────────────────────────────────────────
...
  /api/salon-admin/{salonId}/booking:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    get:
      tags: [AdminBooking]
      summary: List all bookings for a salon (admin)
      description: Admin-only. Returns all bookings across all statuses.
      operationId: listBookings
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /api/salon-admin/{salonId}/booking/{bookingId}:
    parameters:
      - $ref: "#/components/parameters/SalonId"
      - $ref: "#/components/parameters/BookingId"

    get:
      tags: [AdminBooking]
      summary: Get a booking by ID (admin)
      operationId: adminGetBooking
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [AdminBooking]
      summary: Reschedule a booking (admin)
      description: |
        Admin-only. Moves a booking to a different date, time, or staff member. End time is
        recalculated automatically from the service duration.
        Emits **BookingRescheduledEvent** → customer receives a reschedule notification.
      operationId: rescheduleBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingRequest"
      responses:
        "200":
          description: Booking rescheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [AdminBooking]
      summary: Delete a booking (admin)
      description: Admin-only. Permanently removes a booking record.
      operationId: deleteBooking
      responses:
        "204":
          description: Booking deleted
  /api/salon-admin/{salonId}/booking/{bookingId}/confirm:
    parameters:
      - $ref: "#/components/parameters/SalonId"
      - $ref: "#/components/parameters/BookingId"

    post:
      tags: [AdminBooking]
      summary: Confirm a booking (admin)
      description: |
        Admin-only. Moves the booking from PENDING to CONFIRMED.
        Emits **BookingStatusChangedEvent** → customer receives a confirmation notification.
      operationId: confirmBooking
      responses:
        "200":
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/salon-admin/{salonId}/booking/{bookingId}/cancel:
    parameters:
      - $ref: "#/components/parameters/SalonId"
      - $ref: "#/components/parameters/BookingId"

    post:
      tags: [AdminBooking]
      summary: Cancel a booking (admin)
      description: |
        Admin-only. Moves the booking to CANCELLED status.
        Emits **BookingStatusChangedEvent** → customer receives a cancellation notification.
      operationId: cancelBooking
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/salon-admin/{salonId}/booking/{bookingId}/complete:
    parameters:
      - $ref: "#/components/parameters/SalonId"
      - $ref: "#/components/parameters/BookingId"

    post:
      tags: [AdminBooking]
      summary: Mark a booking as completed (admin)
      description: |
        Admin-only. Marks the appointment as COMPLETED after the customer's visit.
        Emits **BookingStatusChangedEvent** → customer receives a completion notification.
      operationId: completeBooking
      responses:
        "200":
          description: Booking completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/salon-admin/{salonId}/booking/{bookingId}/no-show:
    parameters:
      - $ref: "#/components/parameters/SalonId"
      - $ref: "#/components/parameters/BookingId"

    post:
      tags: [AdminBooking]
      summary: Mark a booking as no-show (admin)
      description: |
        Admin-only. Records that the customer did not attend their appointment.
        Emits **BookingStatusChangedEvent** → customer receives a no-show notification.
      operationId: noShowBooking
      responses:
        "200":
          description: Booking marked as no-show
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Customer — Holidays (read-only) ─────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by the identity provider after the admin authenticates.
        **Not yet enforced** — authentication/authorisation is a planned future addition.
        Currently the `salonId` path parameter is the only enforcement in place.
  parameters:
    SalonId:
      name: salonId
      in: path
      required: true
      schema:
        type: string
      description: |
        Salon UUID (`a1b2c3d4-e5f6-7890-abcd-ef1234567890`) or handler slug (`glam-salon`).
        UUID form is tried first; if the value isn't a valid UUID it falls back to a handler
        lookup — the same rule `GET /api/salon/{salonIdOrHandler}` uses. Returns `404` if
        neither matches.
...
    BookingId:
      name: bookingId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Booking ID
  responses:
    NotFound:
      description: Resource not found
  schemas:
...
    AvailableSlot:
      type: object
      properties:
        staffId:
          type: integer
          format: int64
        startTime:
          type: string
          format: time
          example: "10:00"
        endTime:
          type: string
          format: time
          example: "11:00"
        booked:
          type: boolean
          description: >
            true when an existing (non-cancelled) booking already occupies this slot.
            The slot is still returned so the UI can display it as grayed-out / non-selectable.
            false when the slot is open and can be booked.
          example: false
    Booking:
      type: object
      properties:
        id:
          type: integer
          format: int64
        salonId:
          type: string
          format: uuid
        serviceId:
          type: integer
          format: int64
        staffId:
          type: integer
          format: int64
        customerName:
          type: string
        customerEmail:
          type: string
          format: email
        customerPhone:
          type: string
        appointmentDate:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        status:
          $ref: "#/components/schemas/BookingStatus"
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED, NO_SHOW]
    CreateBookingRequest:
      type: object
      required: [serviceId, customerName, customerEmail, appointmentDate, startTime]
      properties:
        serviceId:
          type: integer
          format: int64
        staffId:
          type: integer
          format: int64
          description: Optional; system auto-assigns first available staff if omitted
        customerName:
          type: string
        customerEmail:
          type: string
          format: email
        customerPhone:
          type: string
        appointmentDate:
          type: string
          format: date
        startTime:
          type: string
          format: time
        notes:
          type: string
    UpdateBookingRequest:
      type: object
      properties:
        appointmentDate:
          type: string
          format: date
        startTime:
          type: string
          format: time
        staffId:
          type: integer
          format: int64
        notes:
          type: string

    # ── Website Theme ─────────────────────────────────────────────────────────
...